block-music/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
openai-whisper
numpy